(PC-22191)[PRO] refactor: migrate ProviderForm to tsx, clean tests, remove unused code
