+++ pro/src/pages/Offerers/Offerer/VenueV1/VenueEdition/VenueProvidersManager/AllocineProviderParamaters/AllocineProviderParameters.tsx
@@ -70,7 +70,7 @@ const AllocineProviderParameters = ({
   const initialValues = {
     price: venueProvider.price,
     quantity: venueProvider.quantity,
-    isDuo: venueProvider.isDuo,
+    isDuo: venueProvider.isDuo ?? true,
   }
 
   return (
