+++ pro/src/pages/Offerers/Offerer/VenueV1/VenueEdition/VenueProvidersManager/StocksProviderForm/__specs__/StocksProviderForm.spec.tsx
@@ -1,0 +1,113 @@
+import { screen } from '@testing-library/react'
+import userEvent from '@testing-library/user-event'
+import React from 'react'
+
+import { SynchronizationEvents } from 'core/FirebaseEvents/constants'
+import * as useAnalytics from 'hooks/useAnalytics'
+import { renderWithProviders } from 'utils/renderWithProviders'
+
+import StocksProviderForm, {
+  StocksProviderFormProps,
+} from '../StocksProviderForm'
+
+const mockLogEvent = jest.fn()
+
+const renderStocksProviderForm = (props: StocksProviderFormProps) => {
+  renderWithProviders(<StocksProviderForm {...props} />)
+}
+
+describe('src | StocksProviderForm', () => {
+  let props: StocksProviderFormProps
+
+  beforeEach(() => {
+    props = {
+      offererId: 36,
+      providerId: 2,
+      saveVenueProvider: jest.fn(),
+      siret: '12345678901234',
+      venueId: 1,
+      hasOffererProvider: true,
+    }
+
+    jest.spyOn(useAnalytics, 'default').mockImplementation(() => ({
+      ...jest.requireActual('hooks/useAnalytics'),
+      logEvent: mockLogEvent,
+    }))
+  })
+
+  it('should display an import button', async () => {
+    renderStocksProviderForm(props)
+    // then
+    expect(
+      screen.queryByRole('button', { name: 'Importer les offres' })
+    ).toBeInTheDocument()
+  })
+
+  describe('on form submit', () => {
+    it('should display the spinner while waiting for server response', async () => {
+      // given
+      renderStocksProviderForm(props)
+      const submitButton = screen.getByRole('button', {
+        name: 'Importer les offres',
+      })
+
+      // when
+      await userEvent.click(submitButton)
+
+      // then
+      expect(
+        screen.getByText('Certaines offres ne seront pas synchronisées')
+      ).toBeInTheDocument()
+
+      const confirmImportButton = screen.getByRole('button', {
+        name: 'Continuer',
+      })
+      await userEvent.click(confirmImportButton)
+
+      expect(
+        screen.getByText('Vérification de votre rattachement')
+      ).toBeInTheDocument()
+      expect(props.saveVenueProvider).toHaveBeenCalledWith({
+        providerId: 2,
+        venueId: 1,
+        venueIdAtOfferProvider: '12345678901234',
+      })
+    })
+
+    it('should track on import', async () => {
+      renderStocksProviderForm(props)
+
+      const submitButton = screen.getByRole('button', {
+        name: 'Importer les offres',
+      })
+      await userEvent.click(submitButton)
+
+      expect(mockLogEvent).toHaveBeenCalledTimes(1)
+      expect(mockLogEvent).toHaveBeenNthCalledWith(
+        1,
+        SynchronizationEvents.CLICKED_IMPORT,
+        {
+          offererId: 36,
+          venueId: 1,
+          providerId: 2,
+        }
+      )
+
+      const confirmImportButton = screen.getByRole('button', {
+        name: 'Continuer',
+      })
+      await userEvent.click(confirmImportButton)
+
+      expect(mockLogEvent).toHaveBeenCalledTimes(2)
+      expect(mockLogEvent).toHaveBeenNthCalledWith(
+        2,
+        SynchronizationEvents.CLICKED_VALIDATE_IMPORT,
+        {
+          offererId: 36,
+          venueId: 1,
+          providerId: 2,
+        }
+      )
+    })
+  })
+})
