--- pro/src/pages/Offerers/Offerer/VenueV1/VenueEdition/VenueProvidersManager/VenueProvidersManager.module.scss
@@ -1,21 +1,0 @@
-@use "styles/variables/_colors.scss" as colors;
-@use "styles/mixins/_fonts.scss" as fonts;
-@use "styles/mixins/_rem.scss" as rem;
-
-.venue-providers-manager {
-  .main-list-title {
-    @include fonts.title4;
-
-    border-bottom: thin solid colors.$grey-medium;
-    margin: rem.torem(12px) 0;
-    padding-bottom: rem.torem(8px);
-    display: flex;
-    flex-flow: row wrap;
-
-    .main-list-title-text {
-      @include fonts.title4;
-
-      flex: 1 1;
-    }
-  }
-}
